Fix a warning: size_t to int conversion

+++ tutorial/detection/object/tutorial-detection-object-mbt2.cpp
@@ -145,7 +145,7 @@ int main(int argc, char ** argv) {
         vpHomogeneousMatrix(0.02143385294, 0.1098083886, 0.5127439561, 2.087159614, 1.141775176, -0.4701291124),
         vpHomogeneousMatrix(0.02651282185, -0.03713587374, 0.6873765919, 2.314744454, 0.3492296488, -0.1226054828),
         vpHomogeneousMatrix(0.02965448956, -0.07283091786, 0.7253526051, 2.300529617, -0.4286674806, 0.1788761025)};
-    for(size_t i = 0; i < 3; i++) {
+    for(int i = 0; i < 3; i++) {
       vpImageIo::read(I, imageName[i]);
       if (i==0) {
         display.init(I, 10, 10);
